Added Core Most Gamplay Elements
All data is read in from the JSON file.
A word is spawned after a start delay, and another is spawned when the previous word is clicked on.
When spawned, the word is chosen at random.
When spawned, the word text is sanitised to only contain the correct characters, and be capitalised.
When clicked on, words play the correct sound clip from the mp3 file.

KNOWN ISSUE: Eventually, the audio memory will be overloaded as I currently do not unload audio clips. This will be fixed soon.